requirements.txt
aiohttp
git+https://github.com/hydrogram/hydrogram.git@dev
python-dotenv
tgcrypto
motor
aiofiles
dnspython
pymongo
